Moved adj_jac_apply to rev/mat/functor (Issue #1015)
